Big refactoring
- Created proper packages “commands” and “config”
- Created new makefile target “test”
