+++ commands/commands.go
@@ -1,0 +1,20 @@
+package commands
+
+// This file defines some variables to be available in all commands
+
+var (
+	// token flag
+	cmdToken string
+
+	// verbose flag
+	cmdVerbose bool
+
+	// cluster ID flag
+	cmdClusterID string
+
+	// description flag
+	cmdDescription string
+
+	// TTL (time to live) flag
+	cmdTTLDays int
+)
